requirements.txt
Django==1.8.5
djangorestframework==3.3.0
Pygments==2.0.2
wheel==0.24.0
